+++ zmq/tests/test_auth.py
@@ -9,11 +9,11 @@ import shutil
 import sys
 import tempfile
 
+import pytest
+
 import zmq.auth
-from zmq.auth.ioloop import IOLoopAuthenticator
 from zmq.auth.thread import ThreadAuthenticator
 
-from zmq.eventloop import ioloop, zmqstream
 from zmq.utils.strtypes import u
 from zmq.tests import BaseZMQTestCase, SkipTest, skip_pypy
 
@@ -326,6 +326,11 @@ class TestIOLoopAuthentication(BaseAuthTestCase):
     """Test authentication running in ioloop"""
 
     def setUp(self):
+        try:
+            from tornado import ioloop
+        except ImportError:
+            pytest.skip("Requires tornado")
+        from zmq.eventloop import zmqstream
         self.fail_msg = None
         self.io_loop = ioloop.IOLoop()
         super(TestIOLoopAuthentication, self).setUp()
@@ -335,6 +340,7 @@ class TestIOLoopAuthentication(BaseAuthTestCase):
         self.pullstream = zmqstream.ZMQStream(self.client, self.io_loop)
     
     def make_auth(self):
+        from zmq.auth.ioloop import IOLoopAuthenticator
         return IOLoopAuthenticator(self.context, io_loop=self.io_loop)
 
     def tearDown(self):