+++ zmq/tests/test_ioloop.py
@@ -9,11 +9,12 @@ import pytest
 
 import zmq
 from zmq.tests import BaseZMQTestCase, have_gevent
-from zmq.eventloop import ioloop
 try:
     from tornado.ioloop import IOLoop as BaseIOLoop
+    from zmq.eventloop import ioloop
+    _tornado = True
 except ImportError:
-    from zmq.eventloop.minitornado.ioloop import IOLoop as BaseIOLoop
+    _tornado = False
 
 
 def printer():
@@ -46,6 +47,11 @@ class Delay(threading.Thread):
 
 class TestIOLoop(BaseZMQTestCase):
 
+    def setUp(self):
+        if not _tornado:
+            pytest.skip("tornado required")
+        super(TestIOLoop, self).setUp()
+
     def tearDown(self):
         super(TestIOLoop, self).tearDown()
         BaseIOLoop.clear_current()
@@ -66,28 +72,6 @@ class TestIOLoop(BaseZMQTestCase):
         else:
             self.fail("IOLoop failed to exit")
     
-    def test_poller_events(self):
-        """Tornado poller implementation maps events correctly"""
-        req,rep = self.create_bound_pair(zmq.REQ, zmq.REP)
-        poller = ioloop.ZMQPoller()
-        poller.register(req, ioloop.IOLoop.READ)
-        poller.register(rep, ioloop.IOLoop.READ)
-        events = dict(poller.poll(0))
-        self.assertEqual(events.get(rep), None)
-        self.assertEqual(events.get(req), None)
-        
-        poller.register(req, ioloop.IOLoop.WRITE)
-        poller.register(rep, ioloop.IOLoop.WRITE)
-        events = dict(poller.poll(1))
-        self.assertEqual(events.get(req), ioloop.IOLoop.WRITE)
-        self.assertEqual(events.get(rep), None)
-        
-        poller.register(rep, ioloop.IOLoop.READ)
-        req.send(b'hi')
-        events = dict(poller.poll(1))
-        self.assertEqual(events.get(rep), ioloop.IOLoop.READ)
-        self.assertEqual(events.get(req), None)
-
     def test_instance(self):
         """Green IOLoop.instance returns the right object"""
         loop = ioloop.IOLoop.instance()
@@ -115,7 +99,7 @@ class TestIOLoop(BaseZMQTestCase):
         self.assertEqual(rep.closed, True)
         
 
-if have_gevent:
+if have_gevent and _tornado:
     import zmq.green.eventloop.ioloop as green_ioloop
     
     class TestIOLoopGreen(BaseZMQTestCase):