+++ zmq/tests/test_socket.py
@@ -3,6 +3,7 @@
 # Distributed under the terms of the Modified BSD License.
 
 import copy
+import json
 import os
 import platform
 import time
@@ -235,12 +236,17 @@ class TestSocket(BaseZMQTestCase):
         a.connect(iface)
         time.sleep(0.1)
         p1 = a.send(b'something', copy=False, track=True)
-        self.assertTrue(isinstance(p1, zmq.MessageTracker))
-        self.assertFalse(p1.done)
+        assert isinstance(p1, zmq.MessageTracker)
+        assert p1 is zmq._FINISHED_TRACKER
+        # small message, should start done
+        assert p1.done
+
+        # disable zero-copy threshold
+        a.copy_threshold = 0
+
         p2 = a.send_multipart([b'something', b'else'], copy=False, track=True)
-        self.assert_(isinstance(p2, zmq.MessageTracker))
-        self.assertEqual(p2.done, False)
-        self.assertEqual(p1.done, False)
+        assert isinstance(p2, zmq.MessageTracker)
+        assert not p2.done
 
         b.bind(iface)
         msg = self.recv_multipart(b)
@@ -491,6 +497,39 @@ class TestSocket(BaseZMQTestCase):
         a.send(buf, copy=False)
         rcvd = b.recv()
         assert rcvd == buf
+    
+    def test_custom_serialize(self):
+        a, b = self.create_bound_pair(zmq.DEALER, zmq.ROUTER)
+        def serialize(msg):
+            frames = []
+            frames.extend(msg.get('identities', []))
+            content = json.dumps(msg['content']).encode('utf8')
+            frames.append(content)
+            return frames
+        
+        def deserialize(frames):
+            identities = frames[:-1]
+            content = json.loads(frames[-1].decode('utf8'))
+            return {
+                'identities': identities,
+                'content': content,
+            }
+        
+        msg = {
+            'content': {
+                'a': 5,
+                'b': 'bee',
+            }
+        }
+        a.send_serialized(msg, serialize)
+        recvd = b.recv_serialized(deserialize)
+        assert recvd['content'] == msg['content']
+        assert recvd['identities']
+        # bounce back, tests identities
+        b.send_serialized(recvd, serialize)
+        r2 = a.recv_serialized(deserialize)
+        assert r2['content'] == msg['content']
+        assert not r2['identities']
 
 
 if have_gevent and not windows: